daily build
added better detail to homePage.. adding in sizing batches using srcset

need to resolve max width on a tablet and maybe placing it in landscape-only.

also need to rename images to be more type friendly (long phew)

we'll revisit max-width once we build the wallet page, give us a reference on size and user flow.